refactor: changed names
