+++ src/handlers/compress.js
