+++ src/handlers/hash.js
