+++ src/index.js
@@ -8,20 +8,20 @@ import {
   stdout as output,
 } from 'node:process'
 import * as readline from 'node:readline'
-import handleAdd from './handlers/handleAdd.js'
-import handleCat from './handlers/handleCat.js'
-import handleCd from './handlers/handleCd.js'
-import handleCompress from './handlers/handleCompress.js'
-import handleCp from './handlers/handleCp.js'
-import handleDecompress from './handlers/handleDecompress.js'
-import handleHash from './handlers/handleHash.js'
-import handleLine from './handlers/handleLine.js'
-import handleLs from './handlers/handleLs.js'
-import handleMv from './handlers/handleMv.js'
-import handleOs from './handlers/handleOs.js'
-import handleRm from './handlers/handleRm.js'
-import handleRn from './handlers/handleRn.js'
-import handleUp from './handlers/handleUp.js'
+import handleAdd from './handlers/add.js'
+import handleCat from './handlers/cat.js'
+import handleCd from './handlers/cd.js'
+import handleCompress from './handlers/compress.js'
+import handleCp from './handlers/cp.js'
+import handleDecompress from './handlers/decompress.js'
+import handleHash from './handlers/hash.js'
+import handleLine from './handlers/line.js'
+import handleLs from './handlers/ls.js'
+import handleMv from './handlers/mv.js'
+import handleOs from './handlers/os.js'
+import handleRm from './handlers/rm.js'
+import handleRn from './handlers/rn.js'
+import handleUp from './handlers/up.js'
 import displayCurrentDirectory from './helpers/displayCurrentDirectory.js'
 
 chdir(homedir())
